procstat: fix refcount leak during stats read
Procstat mistakenly assumed that 'forget' will be called immedeately after file
read (instead of close). This was an incorrect assumtion as kernel only calls
'forget' when inode is out of fscache. This lead to refcount leak and resulted in a
fact that stats that were deleted still present in tree.
This can lead to 2 issues:
   1) Prevent from recreating stats with same name
   2) On each "read" from stats refcount will be increased which will result in
refcounter wraparound and release.